HitHighlighter Class: avoid PHP 8 errors
This change should avoid such errors:
PHP Fatal error: Uncaught ArgumentCountError: array_merge() does not accept unknown named parameters

call_user_func_array() expects a function name and an array of parameters to be unpacked and passed on to the function.
In PHP 8, when call_user_func_array() receives an array of parameters with string keys, the keys will be interpreted as parameter names.